test: extract StatusReporterImpl to break coverage cascade
The PR-193 unit test imported ``StatusReporterImpl`` from
``app.views.main_window``, which loads the full QMainWindow assembly
(``preview_pane``, ``layout_manager``, ``image_tasks``,
``dialog_handler``, ...). Those modules then registered in coverage.py
at near-zero coverage and dragged the global total from 88% (master)
down to 65%, tripping ``fail_under = 80``.

Same trap that drove the ``ActionHandlersImpl`` extraction in #182.
Move ``StatusReporterImpl`` into its own module under
``app/views/components/`` so tests can target it directly; main_window
re-imports it from the new location. Behavior unchanged.
