+++ docs/testing.md
@@ -116,6 +116,7 @@ calls out what would be uncaught even with a green CI.
 | `app/views/dialogs/scan_dialog.py` | 90% | every scenario opens it; s17 (full source-list operations); s38 (#144 — inline error when typed path doesn't exist, error clears on next add) | uncovered 10% is QFileDialog browse interaction + a few worker-signal branches. The path-field validation surface (`_on_add_typed` + `_clear_path_error`) is pinned at layer 1 by `TestPathFieldEntry` (empty no-op, non-existent path → visible error label naming the path, keystroke clears stale error, valid path emits + clears prior error) with s38 mirroring end-to-end via UIA against the live dialog — without s38 a regression that broke the QLabel's accessible-name surface would pass layer 1 but the user still wouldn't see the error in the real dialog. |
 | `app/views/components/menu_controller.py` | 89% | s01, s18, s21, s22, s28 | uncovered 11% is fallback branches in the language picker (no available locales) and a defensive guard for missing manifest-actions; the View → Language exclusivity + Yes/No confirm + dirty-flag exit prompt all unit-tested in `test_menu_controller_manifest_actions.py` |
 | `app/views/components/status_messages.py` | 95% | indirectly via every scenario that asserts on status-bar copy (s01, s12, s13, s14, s20, s21, s27, s29) | low — pure formatter; `test_status_messages.py` pins the output shape so qa-explore regexes stay coherent |
+| `app/views/components/status_reporter_impl.py` | 100% | s37 (#138, #140 — baseline) | thin StatusReporter bridge to MainWindow. Extracted from `main_window.py` so unit tests can import it without cascade-loading the QMainWindow assembly (same trap as `action_handlers.py` from #182). |
 | `app/views/dialogs/execute_action_dialog.py` | 83% | s13 (real send2trash through the GUI), s30 (Phase A right-click parity — opens the regex dialog from the Execute tree's context menu), s33 (#166 — banner renders the flagged group number), s34 (#182 — pre-execute lock-confirm Cancel verdict) | uncovered 17% is the actual destructive `_on_execute` flow + a few error branches in the path-not-found dialog; s13 covers the destructive happy path. Spot-add a layer-2 test only if a destructive-flow bug surfaces that's hard to reproduce via the GUI. Lock at execute stage (#182, supersedes #175) pinned by `TestExecuteDialogLock` (single-row + regex flows: each verdict for mixed-locked, all-locked uses the dialog too) and `TestExecuteRequestedLockConfirm` (pre-execute scan: no locked-delete → fast path; APPLY_ALL_UNLOCKED unlocks then executes; APPLY_UNLOCKED_ONLY clears decision on locked + executes the rest; CANCEL aborts). Banner jump-to (#166) pinned at layer 1 by `TestBannerJumpTo`: anchor rendering, `_on_jump_to_group` selects the matching group row, invalid/unknown hrefs are no-ops; the QLabel HTML-anchor click itself isn't UIA-clickable, so s33 only verifies the banner renders the right group number — the click → scrollTo dispatch stays unit-tested. |
 | `app/views/dialogs/locked_rows_confirm_dialog.py` | 100% | s32 (bulk regex trigger), s34 (Execute trigger) | the dialog itself is data + button wiring; `TestLockedRowsConfirmDialog` pins body text shape (count + first-5-basenames + "…and N more"), button-state (Apply Unlocked Only disabled in the all-locked degenerate case), verdict per button click, Esc→Cancel, initial-state→Cancel. |
 | `app/views/layout/layout_manager.py` | 86% | s01 (initial half-screen sizing + adjust-splitter on first manifest load), s39 (#136 splitter min-width floor) | low — the `setup_main_layout` constraints (`setChildrenCollapsible(False)` + `setMinimumWidth(200)` on each child) are pinned by `test_layout_manager_splitter.py`'s splitter-floor tests. Drift would be a removed line, not a behavioural change — visible immediately in CI. |
